% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fct_levels_idx.R
\name{compute_fct_levels_idx}
\alias{compute_fct_levels_idx}
\title{Computing factor level index in respect to another factor}
\usage{
compute_fct_levels_idx(tbl, idx_var, grouping_var)
}
\arguments{
\item{tbl}{input data set}

\item{idx_var}{index variable name (unquoted)}

\item{grouping_var}{grouping variable name (unquoted)}
}
\value{
a tibble with 3 columns:
the first column contains the grouping factor levels,
the second column contains the index variable levels
and the third column contains the computed index
}
\description{
for a given variable,
calculates the index of a level included in this variable
in contrast to groups provided by levels of another variable
}
\examples{
compute_fct_levels_idx(tbl = survey_data, idx_var = s_age, grouping_var = s_gender)
}
