% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_data}
\alias{survey_data}
\title{survey data associated with experiment data}
\format{
A data frame with 892 rows and 100 variables
}
\source{
data provided by Gradient Metrics for Homework Assignment
\url{https://docs.google.com/document/d/1NpkDqtKKxesKislnQuTwsdixwpic-REHmByeAeRFjhs/edit}
#' @usage data(survey_data)
}
\usage{
survey_data
}
\description{
data from a regular survey with a set of questions related to
sleep and socio-demographic behavior
}
\keyword{datasets}
